[update] リファクタリングをしました
コンポーネントのファイル構成を一新しました
