+++ app/(home)/[id]/_components/ContainerSkeleton.tsx
@@ -1,11 +1,11 @@
 import { Skeleton } from "@/components/ui/skeleton";
-import { Card } from "../ui/card";
+import { Card } from "@/components/ui/card";
 import React from "react";
 
 const ContainerSkeleton = () => {
   return (
     <Card>
-        <Skeleton className="w-full h-screen"/>
+      <Skeleton className="w-full h-screen" />
     </Card>
   );
 };
